Refactored the runtime and container API
Finally added back quick-launching of runtimes: the new implementation works
in a generic way across any container and runtime combination, as long as at
least the container supports quick-launching in the first place.
The only thing missing is the re-integration of Pelagicore's "pelagicontain"
LXC integration, but this will be added in a later commit, preferably via a
new plugin interface.

Change-Id: I8ba46b21b7123450dea5fc8bac824518abf2e51f
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>