Fix race condition in DBus registration/connections.
When an app connects to the AM's P2P DBus, it immediately connects to the
interfaces via QDBusInterface(). The problem is that the server is
registering those interfaces at the same time, leading to race conditions.
Somehow these race conditions are only showing up now in newer 5.7
builds, although the problem has been there for a long time.

The actual solution is not very pretty, but gets the job done. If there
is a nicer way to accomplish the trick, we should use it.

Change-Id: Ieaac52f4ada6f6a25f8b8717ed8fddd44fd30503
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>