Take aliases into account when starting apps via openUrl.
Instead of starting the first matching app entry for a given url scheme, try
harder to match the rest of the url against the documentUrl fields in the
info.yaml files. In case we have a match, start the app via the alias entry,
otherwise always use the non-aliases entry.

Change-Id: Id458360a22abac8961b84e023111f81c2bd535bd
Reviewed-by: Nedim Hadzic <nedim.hadzic@pelagicore.com>