Doc: Convert doc config to use standard Qt module make targets
With this as a submodule of a module that uses the stardard
Qt documentation targets (prepare_docs, generate_docs, etc.),
the build was failing. Fix this by adopting the same
documentation targets as a Qt module.

This allows doc builds of applicationmanager as part of a larger
project. Also, because there is now a prepare_docs step that
generates .index files, this enables proper cross-module linking
via doc dependendencies.

Task-number: AUTOSUITE-8
Change-Id: I5db9365faf34ca5ef7c542792b77b9b1a1b571bc
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>