Improve DLT logging
Use the new lazy initialization feature of QtGeniviExtras.
As the application-manager has it's own logging-category parsing the
log levels of the categories might change after they are defined. As
DLT doesn't support it, the lazy initalisation is used to make sure
the DLT-Viewer is showing the correct state of the logging categories

Change-Id: I7985b86fb24699b3b792fef242351d8114cd5c4a
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>