Reactivated the -config install-prefix option
If you are not specifying an install-prefix when running qmake, then the AM
will behave like a standard Qt module:
 - in developer builds, the binaries will be compiled directly to $qtbase/bin
 - in prefix builds, the binaries will be compiled to $builddir/bin

What's new is the behavior when an install-prefix has been given:
 - in developer builds, the binaries will be compiled directly to $install-prefix/bin
 - in prefix builds, the binaries will be compiled to $builddir/bin

In addition, all binaries will get an absolute R(UN)PATH when configured with
an install-prefix. This will allow you to run the binaries from the
build directory directly without weird side-effects (Linux will just pull in
the system-Qt if it cannot find the libs in R(UN)PATH.

In addition, the D-Bus XML files are not installed by default anymore, since there's
no fitting installation location within Qt.

Change-Id: Ic81389c340f88f710f129407876f304af859b0e1
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>