Emit application specific memory warnings
When an application (process) has been moved into a memory controlling
cgroup and the application exceeded its allotted memory, it simply got
killed by the kernel. This is still the case, but the application will
additionally get informed beforehand when memory is getting low and
critical. The thresholds for these warnings are currently predefined
to default values (75/90%).
Fixed system wide memory low and critical warnings, as well.

Change-Id: I2426c04a4ed5958a35ac07fb70e6ed7c2bb83ead
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>