Make ApplicationInterfaceExtension object creation dynamic
If we use preload apps or use --single-app parameter, application
interfaces will not be created in the main QML and an app won't be
able to use those interfaces. We add the connection when new
application interface is created so the client side is informed.
This allows us to create interface extensions during the runtime
and on demand.

Change-Id: I4af950db9cea043706fd12a90cdc98dd629b0008
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>