Add systemProperties to configuration
The new field "systemProperties" will eventually replace
"additionalConfiguration" in config.yaml files. Backwards compatibility
will be kept for a certain transition period.
This new field allows to expose properties to the system UI, built-in
and 3rd party apps in a more specific manner.

This is also a prepartion for the introduction of "applicationProperties"
(defined in info.yaml files) in a subsequent patch.

Change-Id: Ifcb371debbb1320f717e51956cc58e0736d36924
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>