Improving the mimeType API
In openUrl() we now check for the full documentUrl when searching
for an matching alias to open. Previously we removed the schema
before, which could lead to problems if the application can handle
multiple schemas.

If openUrl() is used with a file not known to any application
and the mimeType is used, the detected mimeType is now also
forwarded to the application.

Change-Id: I6af8bd5459e6ce34745af0c3f5027b18d6664ab2
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>