Addition of a qmlplugindump tool equivalent for the application-manager
The AM is not built as a plugin that can be introspected by the qmlplugindump
tool. This information is necessary though to support auto-completion in
QtCreator.
The new tool to achieve this is appman-dumpqmltypes - start with --help to
get more information.

Some code had to be shuffeled around to make all QML types available in
libs that can be linked into the dumper tool.

Change-Id: I0d934e71edb97bca4bdfd6893b49acca66ceac11
Reviewed-by: Thomas Senyk <thomas.senyk@pelagicore.com>