Refactored the old and the new Wayland compositor into one .h/.cpp file
We get a lot more #ifdefs this way, but at least nobody can forget the old
compositor that easily now when doing changes to the current one. Also, those
ifdefs should be gone when we can drop 5.6 support (hopefully when 5.9 is
released).

As an added bonus, the compositor is now working again with Qt 5.6.

Change-Id: I456cc211aca26d2f17c4466568d7dd620779ed63
Reviewed-by: Thomas Senyk <thomas.senyk@pelagicore.com>