Introduce config file caching
After parsing all config files specified on the command line, the appman
will write a binary dump of the resulting configuration QVariantMap to a
cache file via QDataStream. This cache file includes absolute paths and
checksums of all config files, so that the appman can verify if the
cache is still up-to-date on subsequent starts.

Change-Id: I1d6f6621f75d1bba7849b3514177fc2c245c4c23
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>