Improve startup timing
Startup time reports were not very useful for quick-launched
applications since they included the leap between the process pre-fork
and the actual user start. Besides there is generally no value in
measuring process creation running in the background.
The timer will be reset in this case to the actual start of the
application.

Also fixed output of title when report was requested a second time and
checkpoint order.

Change-Id: I8690185f8199684a43655a4ad45ce5152aed72cf
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>