Replace chapter with section1 in documentation
Chapters are printed on the same hierarchical level as section1, so they
could not be used to give hierarchical structure. For consistency
chapter has been replaced with section1 in the entire documentation.
Sub-sections can be defined with section2, section3 and so on. This was
relevant for the "Logging and Debugging" documentation in particular.
Corrected some other mistakes, as well.

Change-Id: I3e191c96c802fd906b1dc1085bc1c6d8fe68fda8
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>