Allow to show root window in single-process mode
In single-process mode, if the initial visible property of a
FakeApplicationManagerWindow root objects has been false, changing it
later would not cause it to appear (no windowReady signal is
emitted). The reason was that the root FakeApplicationManagerWindow
has no QObject parent. The check for a parent has been removed now.

Also added an auto-test for that and a test that reveals more
differences between single- and multi-process mode.

Task-number: AUTOSUITE-131
Change-Id: Icec4c5f4ebe1c0ccf76e11fcb0ac3a82ff1ec335
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>