Hide properties in FakeApplicationManagerWindow
To mitigate the difference between single- and multi-process mode,
properties and signals in FakeApplicationManagerWindow (Item) will be
hidden, that are not available in ApplicationManagerWindow (Window).
The same warnings will be printed and applications will not start,
which use such properties. This patch only hides "parent" - it
provides a base for further properties that will be hidden.
To achieve this an InProcessSurfaceItem has been introduced that
is actually exposed to the systemUI and serves as an item parent for
FakeApplicationManagerWindows.
Also changed the lifetime of the application's object hierarchy.

Task-number: AUTOSUITE-143
Change-Id: Ib4881ec39267907fb24bebbd167796270661ab8c
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>