Hide more properties in FakeApplicationManagerWindow
A decission had to be made whether the logging output (warnings)
should closely resemble the ones in multi-process mode or whether the
termination on error behavior should be the same. The latter has
been favored. For instance binding a none-existent property leads to
an abort, whereas assignment doesn't.

Task-number: AUTOSUITE-143
Change-Id: Id7b0f6eea77236855066281835ec6293ccc4850c
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>