Fix the version comparison for enabling the dlt lazy init
With Qt version 5.9 no MAJOR_, MINOR_ and PATCH_VERSION variables are
generated for the modules anymore.

Because of that the lazy initialization of the DLT categories were
never enabled, which caused appman to only register the Categories
which are used for logging instead of all of them.

A generic qmake test function is now used for the version comparison of
the VERSION string.

In newer Qt version (5.10.0) qmake provides a builtin function

Task-number: AUTOSUITE-169
Change-Id: I76b40320167835b10b159ec221a100f3b1276f88
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>