Add a deterministic way to start apps on startup
Apps that are started before the compositor is ready to accept windows are
now delayed until the new (public) property
ApplicationManager::windowManagerCompositorReady
signals that it is safe to do so.

Also, WindowManager::registerCompositorView() can now be called explicitly on
the System-UI's root window, so that the registration is done earlier (the AM
will always implicitly register that window after the main QML is loaded).

Change-Id: Id461af2dcc64258c4fab73a87a4f55aa32ab3654
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>