Make the minidesk and monitor examples launchable from qt-creator
For project like these there is now a new qmake feature called
am-systemui. This feature will copy all files and folder defined
by AM_COPY_DIRECTORIES and AM_COPY_FILES to the build-folder and
create a runner script which executes appman with the the arguments
defined by AM_DEFAULT_ARGS.
This script is setup as the qmake TARGET and can be started by
qt-creator.

This trick has the limitation that debugging is not supported

Change-Id: If93d7971df646163e9819a3c5e8a436fc76bef21
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>