Refactor the Application object
Split the original Application object into two parts:
the dynamic and the static one.
The dynamic is called Application and lives in the manager lib
The static is called ApplicationInfo and lives in the application lib

On top of that make a clear distinction between aliases and real application
objects, which solves some inconsistencies.

Also removed all deprecated properties from the original Application object

Task-number: AUTOSUITE-296
Change-Id: I8354c9d50f14d083cb85c19153dd19c933eae660
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>