Rewrite the tst_installer QML autotest
Instead of relying on signal spys, we now use the javascript connect
method to directly connect slots which do the state comparison the
moment the state has changed and save the order of the state changes
to a list.

The order is then checked after the installation/update is finished.

This fixes also a small problem in the applicationmanager which caused
the application state property to be updated to BeingInstalled before
someone could connect to the statechange in QML.

Change-Id: Ifa25eab6c97520de34247bc8d7c79bb01ca19c68
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>