WindowManager: remove unimplemented IsFullscreen role
That way we can remove quite a bit of code through several classes.
Having a signal specific to inprocess Qml in the AbstractRuntime doesn't
look good anyway.

And the documentation is wrong. It describes a desire of the application
to have its window fullscreen. What the window manager will do with this
information/window state is another story.

Change-Id: If492ba8e81c2a87be20f34b51c2ac4ed19b69bc4
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>