WindowManager model contains only windows with surfaces
This makes it possible to have a Repeater using WindowManager
as a model without having to add any imperative code to unset
WindowItem.window in order to free up the WindowObject once it
loses its surface.

Code that wants to animate a WindowObject destruction has to
keep its WindowObjects in a separate model, where it can
remove them once finished with the animation, which is a more
advanced use case.

Change-Id: If81e90c9e71b1a75c902a73fb6f1d19f0220335d
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>