Update documentation of the minidesk example and some code tweaks
The documentation was still referring to the old API.

Code changes:
* Removed the close button from pop-ups as the WindowObject.close()
API is already being shown in the delegate for regular windows. This
makes minidesk more "mini".
* The application icon can now be used for stopping applications as well.
Otherwise the "Application Termination" wouldn't make sense as there
would be no way to call the Application.stop() method.
* Renamed the id of a delegate to better reflect its content. Doesn't
have just the window chrome but also the window itself.

Change-Id: I6c870c06f62851071495506bd9faad1ade24bfd7
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>