Add DLT crash logging and QML stack trace
When the System-UI or an application crashes we try to get a QML
stack trace in addition to the (C++) backtrace. Both traces will be
written to DLT if enabled, in addition to the terminal output.
Crash tests have been added, as well.

This is a backport of 1feaa3e from the dev branch.

Change-Id: If7b3722ae3d54d3e219c999afc81006abce316c0
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>