Fix the /tmp directory being chmod'ed to 0700 when running with sudo
QStandardPaths will chmod the $XDG_RUNTIME_DIR to 0700, if the user running
the binary is the owner of that directory. I think this is wrong, but Thiago
says "this is what the XDG specs want you to do" (no point in arguing any
further there).
So instead of setting this variable to /tmp (which is always owned by root,
hence the problems with sudo), we do it the "right way" and use a LFS
compliant, user-specific runtime directory.

Change-Id: I5ac5223ebca9bd55f3b049939c7f5aa4a87372a7
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>