Intents: add timeout/error handling for the client side
Right now, both incoming and outgoing request are expected to be answered
within 5sec. This needs to be made configurable in a follow-up patch, since
these values are very system specific.

In addition, the IntentRequest objects now have JS ownership, so that we
can easily react on incoming request to applications not being handled.

Change-Id: Id0703e4efd6f63a2404cde6d658b035f8b0985c1
Reviewed-by: Daniel d'Andrada <daniel.dandrada@luxoft.com>