Make deducing the single-process app context easier
Consolidate our efforts of detecting which single-process app an object
belongs to (or from which app a call into a singleton was triggered) in
QmlInProcessRuntime instead of duplicating this all over this place.
(Notifications are also using a similar method, and need to be switched
over to this scheme in the future).

Change-Id: Iebbb9a8540b3d941ee4af0c661e2f19da7bc6f3e
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>