Stop SystemMonitor updates when all reporting is disabled
When all reports are disabled, stop updating the model
since there's no work to be done. Once reporting is
enabled back again, reset the model and restart the update
timer.

The previous behavior would keep the update timer running,
filling the model with blank data row-by-row until the model
was completely empty, at which point it would finally stop the
update timer. That was a waste of resources and caused many
unnecessary wake ups.

Task-number: AUTOSUITE-688
Change-Id: I41a56b0cab7516b2e4545ca5faeae84c8c015d0d
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>