Get rid of the last remaining direct moc include in source files
Moving those private classes out to a _p.h is cleaner and matches the rest of
the AM code base. It also helps with code-coverage tools that do not like to
have #included source code in the build directory.

Also renamed launchers/qml to tools/launcher-qml. This should have been done
a long time ago and this also makes the code-coverage reports a bit easier to
read (sort order wise).

Change-Id: I70443a0998f1aa2958a7cb47b57824fd8dfae81f
Reviewed-by: Daniel d'Andrada <daniel.dandrada@luxoft.com>