Fixing more UTF-8 problems on macOS: runtime errors in libarchive
When running auto-tests or appman tools in a shell with
LC_CTYPE/LC_ALL not being set or being set to a non-UTF-8 value, libarchive
will not be able to read a from an archive with UTF-8 encoded names.
We're now using the same code path as on Linux to detect and workaround these
situations.

In addition, a DeploymentWarning is generated if the default locale is not
suitable.

(also needed to fix the custom appman example and its docs)

Change-Id: Ib2936fafead518a66efffc2e7bedceb7c46836d3
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>