Add support for isSingleton and isCreatable in the qmltypes type information
Extended the AM-QmlType class info to support trailing tags: at the moment
SINGLETON and UNCREATABLE are supported.

Also added rules to ignore functions and properties starting with '_', since
these are considered to be private in the QML world.

Improved the error handling/dealing with invalid class infos while at it.

Change-Id: Ia68136fe30404e58d808be3a6a04d5f3b9c974c9
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>