Avoid conflicts on Linux desktops by using a non-standard Wayland socket
If not specified differently via command line option or $WAYLAND_DISPLAY,
the AM would just rely on libwayland to pick a name for its Wayland socket.
The problem here is that some desktop environments would see "wayland-0"
(the first compositor registered) and think that they are running in
Wayland mode and consequentially start client applications on this
compositor.

The default for the AM is now "qtam-wayland-0".

Change-Id: I508ac8fb051735dc64de35ea8dfc44491786c7d7
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>