Untangle the broken monitoring dependencies
After the last refactoring of the monitoring components, we ended up with
a monitor_lib that pulled in nearly all of the other AM components. The end
result was that the qml runtime had all the appman system-ui code compiled in
as well.

Downgraded the monitor_lib to just read values from the system and moved the
QML API classes to shared_main_lib, manager_lib and main_lib respectively.

Change-Id: Ic205153c6d8c221793b0f2546b2f31f4e1ede53d
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>