Enable hardware OpenGL acceleration for Wayland clients
When running the AM on an X11 desktop, the Wayland clients will use the
LLVM-pipe software renderer. Setting $QT_WAYLAND_CLIENT_BUFFER_INTEGRATION
to "linux-dmabuf-unstable-v1" will enable the new buffer integration that
will allow full GL acceleration.

We need to make sure however to NOT enable this setting when running with the
propiertary NVIDIA driver, which does not support the linux-dmabuf interface.

See QTBUG-75575 for technical discussion around this.

Change-Id: I7e4bb9ec38ee3eefa359535fc12bf38704c30358
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>