Only calculate the hardwareId once
If no other configuration was chosen appman calculates its hardwareId
using the available network interfaces.
Instead of iterating and sorting them every time the hardwareId is
now saved once calculated.
This is also a workaround for QTBUG-76408 on macOS which reports
a wrong intereface type on the second run and cause the appman to
stop without hardwareId.

Fixes: AUTOSUITE-1019
Change-Id: I521245987486b09fce5c410918d0a35be37b1e10
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>