Fix app start/stop in single-process mode
If an app has been started and immediately stopped again, the next
time the app is started the assert would fail, that checks that the
app's root object is null. The stop signal would be handled before
the InProcessWindow could be created and hence the app's window
surface would never be released. The assert is in
QmlInProcessRuntime::start().

Cherry-picked from dev: cada823

Change-Id: I0711416a0fa4582bdaaa0a7f9b0c081eb6539c01
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>