Add new package abstraction, part 2
The first commit in this series was b4aee167d3bc6b9f64229317fbc428b3f3b83c0d.
This update fixes all the known bugs introduced in the change of the
architecture, adds basic documentation and examples.

The changes in the intent sub-system are in a separate patch for easier review,
although they both depend on each other due to changes in the API.

Change-Id: Ia78da7e10e2eb7adfa30625554af275453065a90
Reviewed-by: Dominik Holland <dominik.holland@qt.io>