Fix canceling installations
Canceling an update left the app in an inconsistent state. There is
still a known issues: canceling the update of an already updated
built-in app  will revert back to the original app. This will be
fixed in 5.14

Also downgrading a built-in app didn't work as expected (the app would
be blocked, the icon etc. wouldn't be updated).

Change-Id: I6c72e7f87e993a6839e97ceb6cea443f07e1fd77
Reviewed-by: Robert Griebl <robert.griebl@pelagicore.com>