Add a Windows crash handler for MSVC And MinGW
Backtraces are not supported on MinGW at the moment, because StackWalker
doesn't compile due to a missing dbghelp import library.

Also improved the async-safety of output functions. On the Linux side, I
noticed a lot of dead-locks and crashes while generating the backtrace lately.

In addition, the getOutputInformation kitchen-sink function has been
split up a bit and moved to the logging.cpp where it logically belongs.

Change-Id: I354c0e4e212a247338faddb6e15d8d0831852086
Reviewed-by: Dominik Holland <dominik.holland@qt.io>