Really fix the installation problem via the appman-controller
The bug was trickier than anticipated, because I thought that only the appman
side would check the meta-data. Turns out the controller tries to verify the
packageId as well, but the meta-data was just not sent over DBus. Why? Because
the Package object may not have been in model yet (on first installation), so
calling PackageManager::get(<id>) led to an empty (error) repsonse.

Also fixed a few left missing s/application/package/ replacements.

Change-Id: I6fc436270ad193203fca94e66e739ffd4baeae27
Fixes: AUTOSUITE-1461
Reviewed-by: Egor Nemtsev <enemtsev@luxoft.com>