Make window order predictable
In-process "windows" need to wait for their QML onCompleted handlers
to be called before they can be added to the System UI to make sure
that their window properties are available. If multiple windows are
created, the order would depend on the existence of onCompleted
handlers. The order has been changed to always match the order of
their componentComplete() function calls. This happens (but is not
guaranteed) to be the same order as in multi-process mode.

Change-Id: I1431277f6fa160b1713e94fc27195b178377e98a
Reviewed-by: Robert Griebl <robert.griebl@qt.io>