appman-controller: Improve option and help usage
The controller accepts the options for the command now also before
the actual command e.g. -ioe debug-application
This makes it easier for the user, as this is also the format printed
by QCommandLineParser::showHelp()

Showing the full help for all commands also when only the specific
help for a command should be provided, makes it very hard to see what
the correct usage for the command actually is.
We now show only a short help when a command is provided and the full
help in all other cases.

Change-Id: I6927cf465ba1dd46cfbf361af04a3bfc8d089b67
Reviewed-by: Robert Griebl <robert.griebl@qt.io>