Fix use-after-free on xdg pong
With the xdg shell, ping is not related to a particular surface.
WaylandCompositor associates the pong with its original WindowSurface
with a map of the serial number to WindowSurface.

If the WindowSurface is destroyed before pong arrives, the map will have
an invalid pointer leading to a compositor crash.

Change-Id: Ide32ac80d42bf8373e8180e12ab89b29e6d00341
Reviewed-by: Robert Griebl <robert.griebl@qt.io>
(cherry picked from commit 7038e1681863c8a59c6c6a758fc7ec41f29f2be7)
Reviewed-by: Bernd Weimer <bernd.weimer@pelagicore.com>