Implement cache typing and versioning
While the new cache architecture has a type and version field for the global
header, I totally forgot to add the same thing to mark the type and version
of the actual cached data.
There was a version field for Package/Intent/ApplicationInfo, but this was
badly implemented and could crash the system trying to load an incompatible
cache file, because the version check was done too late.

Change-Id: I755cc5101213b60996541ea7e872e3017550ec37
Reviewed-by: Dominik Holland <dominik.holland@qt.io>