Retain /proc/<pid>/environ output
While setting the process title provides nice output for the ps
command, it spoils the /proc/<pid>/environ output, which is also
useful to see sometimes.
Added another method for processes under appman control
(appman-launcher-qml), that overwrites a dummy argument. This limits
the additional application Id output to 64 characters currently.
For the sudo helper no dummy argument can be added, but the
environment can still be seen in the parent process.

Change-Id: I2d6bf2e24694ae4bd699f0bea87ad8eb2f51c140
Reviewed-by: Robert Griebl <robert.griebl@qt.io>