Fix potential crash when a window is released
There can be a race condition when a window is released. A crash
happens very rarely in the windowitem2 QML test case (~1%), but could
be seen more often in more complex code.
Because, both window->deleteLater() and isBeingDisplayedChanged are
processed asynchronously, the former might run before the latter and
hence the window might already be destructed when used in the
isBeingDisplayedChanged handler.

Cherry-picked from 5.15: 520d102

Change-Id: Iddb687f057666ef3d34feb6472d3ba2171093d91
Reviewed-by: Robert Griebl <robert.griebl@qt.io>