Fix window properties (show-hide-show)
With QtWayand patch e18cfb5 we cannot rely any more on the fake
PlatformSurface SurfaceCreated event for sending window properties
from client to server, when the window is made visible again. We need
to attach to the Expose event instead. To reduce IPC, redundant
Expose events are ignored. Unfortunately, the Show event cannot be
used since there might not be a mapping from the window to the
Wayland surface.

Change-Id: I476fd6ae2968f90f25aef6533e6eb5c3b26a613b
Reviewed-by: Robert Griebl <robert.griebl@qt.io>