cmake: Fix qml-only tests/example wrapper generation
Use the binary from the build folder inside the wrapper
instead of the installed binary location.

Make sure the wrapper is regenerated correctly when the
CMake variables change.

Set PATH and QT_PLUGIN_PATH in the wrapper scripts similar
to the autotest cmake wrappers.
This should make sure to always use the correct executables.

Change-Id: I318ed549d419363e5e4b194f715f067575cb8596
Reviewed-by: Robert Griebl <robert.griebl@qt.io>