Fix glibc compatibility problem
glibc 2.32 deprecated sys_siglist: the symbol is not available to link
against anymore, but it is available at runtime for apps that have been
linked against an older glibc version.

Our problem is now that the CI builds release version of the AM against
an older glibc, which works for the appman binary itself, but if you
want to compile a custom appman executable on a modern, glibc 2.32
based system, you get a "missing symbol" linker error in
libQt6AppManCommon.a(unixsignalhandler.cpp.o)

Detecting this at runtime would be a nightmare for little gain, so we
fall back to using the non-async-signal-safe strsignal() function
when initially building against a glibc < 2.32.

Change-Id: I1e702c5ee00e2725fe6f76f9ce001c099ec72eea
Fixes: AUTOSUITE-1667
Reviewed-by: Dominik Holland <dominik.holland@qt.io>
(cherry picked from commit 65ddbe72001cc9e5a49b54d821eafd352c25ee92)
Reviewed-by: Robert Griebl <robert.griebl@qt.io>