Rename main.cpp generated by qt6_am_add_systemui_wrapper in cmake examples
This is needed as our doc build will otherwise pick up the generated
main.cpp as an important source file for the examples-manifest.xml
But because the referenced main.cpp doesn't exist in the source tree
QtCreator will complain when opening such an example.

Renaming the file to wrapper.cpp fixes it as qdoc doesn't pick that up.

Fixes: QTBUG-98797
Pick-to: 6.2 6.2.2
Change-Id: Id6b20ae7c113b2bb8dd38fb9e4f8c5ddcb44f4f2
Reviewed-by: Robert Griebl <robert.griebl@qt.io>