Untangle the icon/name/description/categories mess
The documentation and the implementation disagreed quite heavily in
regards to the icon, name, description and categories fields in the
manifest for the package, each application and each intent.

The implementation now follows the existing documentation, but also
relaxing the requirements of icon and name being required for packages.

Also, it was confusing that applications could not override these
properties from the package like intents can do, so they also got
these properties.

Finally, everything is now correctly exposed to QML, with legacy
wrappers in place to not break old code.

In addition, the completely outdated applicationinfo unit-test
was rewritten to test all the current field combinations.

Change-Id: Ie5ad84bc6b771d6702b1994d1c8ae9f0204643b9
Pick-to: 6.3 6.3.0
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>
Reviewed-by: Dominik Holland <dominik.holland@qt.io>