Handle an empty or invalid documentDir correctly
Not specifying the documentDir, but enabling the installer would
- unexpectedly - use $PWD as documentDir. The code does now
follow the documentation: if not set, no document directories
will be created on package installation.

Also, we now fail early if the documentDir and installationDir
are sub-directories of each other, as that would lead to mayhem
on PackageManager::cleanupBrokenInstallations().

Change-Id: Id8d89f82bf0d63c771f3d86a963c66681f43f195
Fixes: QTBUG-101881
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>
(cherry picked from commit bc06a04dd845d039ba802f94b5eae57519ec63a8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>