Correctly track redirected std. IO file descriptors through all layers
Using redirected IO via the DBus startApplication interface could lead
to blocked terminals, due to the forwarded file descriptors not being
closed correctly when errors occurred during app startup.

We're now using move semantics plus a helper function to make sure that
the fd vector always has exactly one owner and that owner is actually
closing the descriptors.

Changing the public ContainerInterface would be an API break, so fixing
this as well is not an option at this time. Meanwhile the documentation
clearly states the required ownership transfer.

Also added a note that IO redirections do not work for single-process
runtimes, when trying to do so via the appman-controller.

Change-Id: Id4fa0e53ee61c33a630ca4a8d6d89aea78cc74e7
Fixes: AUTOSUITE-1218
Reviewed-by: Dominik Holland <dominik.holland@qt.io>