macos: Fix the appman-qmltestrunner
We cannot rely on the QTestRootObject::instance() method
as this function is inlined by the compiler and because
of that doesn't provide the same instance as used in QML.

Instead we use the new singletonInstance() method to
retrieve the instance from the QQmlEngine.

This also enables the qml tests on macos again.

Pick-to: 6.2 6.3
Fixes: QTBUG-102736
Change-Id: Ia73bb9ce54236a30683c1fa9f5fd51a1e456aad4
Reviewed-by: Robert Griebl <robert.griebl@qt.io>