Fix DBus error handling when set to "auto"
Problem 1: A Qt plugin calls QDBusConnection::sessionBus() before we
fork off our own, private session bus. In this case, Qt caches the old
bus address and we try to register our objects on the wrong bus.

 -> register on our own bus explicitly.

Problem 2: If "Problem 1" happened, every QML application would block
for 100 * ~25sec on startup, trying to call "Introspect" on the non-
existing Notification interface (on the wrong dbus). The delay loop
was written for the P2P case, which doesn't block in the QDBusInterface
constructor, resulting in a more reasonable 100 * 1msec timeout.

 -> check if the service is available at all first and also check
    for time-elapsed instead of trying 100 times.

Change-Id: I34ade2bee2da27753eda09b6c0f3562882f40bc3
Reviewed-by: Dominik Holland <dominik.holland@qt.io>