Update commercial license headers
Updated header.COMM to the files in tqtc-qtapplicationmanager.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4946
Change-Id: Iad1d7d910470d49db56f405640d7dd35799ab049
Reviewed-by: Robert Griebl <robert.griebl@qt.io>