Fix autotests to work on windows
The test doesn't work if debug_and_release is set for the app lib
as the libs are created in subfolders in this case. Because of
the special coin test setup the pro files need to be split up.

The simple QML autotest doesn't need special handling on
windows anymore.

Change-Id: I73e0af8483438fa1ff249156d9f6e9f63eb229db
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Robert Griebl <robert.griebl@qt.io>