Support QNX (unoffically)
Allow to build for QNX and run in single-process mode. There are a
few limitations, e.g. filesystem mount monitoring and startup timer
are not supported and there's only a rudimentary stack trace in case
of a crash, etc.

Since Qt Network doesn't provide a HW address and there's currently
a cross-compilation issue, the following options need to be passed to
configure:
-DINPUT_hardware_id=dummyHwId -DQT_FORCE_BUILD_TOOLS=ON

Change-Id: Ia58d09d15a781666fe788f21dfe9c7af7a2a74ef
Pick-to: 6.4
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Robert Griebl <robert.griebl@qt.io>