CMake: our tools are really apps as far as the build system is concerned
The tool syntax got ported over from the old qmake system, but this
leads to problems nowadays on cross or multi-config builds, where the
tools are not built at all, unless they are built explicitly.

What we call tools, really are apps for Qt's cmake build system, so
all of those (except the packager and uploader (*), which really are build-
time tools) have been converted to be built as apps.

(*) the uploader should be retired or moved to the
qt-apps/qtauto-deployment-server repository anyway.

Change-Id: I7be0ce63625e53d0bc65387c5e5200c71c00b48a
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Dominik Holland <dominik.holland@qt.io>