Intents: add broadcasts
This commit adds a new intent type: a broadcast.
Broadcasts will be delivered to all applications (*), but they do not have
the possibility to return a reply back to the sender.

Implementation wise, they are just treated as normal requests that are
multiplexed for every application.

(*) This can be limited with the new handleOnlyWhenRunning flag.

Change-Id: If9f954cf5e52707624b95c80c8e984dfd6c4315a
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Dominik Holland <dominik.holland@qt.io>