Fix a potential crash when trying to start applications
If you were calling startApplication() and the runtime failed to
execute the app's binary in a container, it would call deleteLater() on
the failed runtime object. The problem was that if you immediately
called startApplication() once more in the same event loop iteration,
the Application object still had the to-be-deleted runtime attached,
which in turn prevents the allocation of a new container, so the
AM tried to start the application again, but this time with
"container" set to nullptr.

The actual fix is to NOT call deleteLater, but directly delete the
runtime object.

In addition, when startApplication() is called for an application,
we have to make sure that the application's runtime object can not
be in the state NotRunning: the AM could still end up in this state,
if a Container plugin is either not reporting shutdown correctly, or
if the container's "ready" state is delayed.

Change-Id: Ied1baec8c90d4e0a980c296cbc7cd87b12629524
Pick-to: 6.4 6.3 6.2 5.15
Reviewed-by: Dominik Holland <dominik.holland@qt.io>