Forward keyboard focus to applications
Key events were not passed to applications, when the focus moved to
a WindowItem in the System UI, since the item representing the
application window is actually a child of the WindowItem.

A possible solution would be to explicitly send the key events to the
application item, though the approach taken here is to forward the
keyboard focus. This results in key events always reaching the
application first (even when the System UI would actually consume the
events) and in single-process mode, if the application consumes the
event the System UI won't get it any more.

Fixes: QTBUG-108837
Change-Id: I1699556f1a67057e8be2848e37c846c2ee33a646
Reviewed-by: Robert Griebl <robert.griebl@qt.io>