CMake: Fix missing waylandwindow.h include error
The header stopped being generated by syncqt/cpp on certain
platforms because we only generate headers that meet
the qt_internal_extend_target conditions. In this case qtwayland is
not available on macOS, and thus the waylandwindow.h header is not
generated.

Fix windowframetimer.cpp to include it conditionally only if the
multi process feature is enabled.

Not generating the header is fine in this case, because the whole
module is Private, even though the header has no _p.h suffix.

This unblocks the build with syncqt/cpp.

Fixes: QTBUG-109337
Change-Id: I0a38af6d23e5d1c9e040281a52a3b719106382b9
Reviewed-by: Zoltan Gera <zoltan.gera@qt.io>
Reviewed-by: Robert Griebl <robert.griebl@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>